+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -128,7 +128,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kcg9y1raq8y8jjpgdewp7wen"),
+  _id: z.string().default("conv_01kchggrre54051zbypkx0j0td"),
   entityId: z.string(),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
@@ -161,7 +161,7 @@ export const ConversationResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationResponse
 > = z.object({
-  id: z.string().default("conv_01kcg9y1raq8y8jjpgdewp7wen"),
+  id: z.string().default("conv_01kchggrre54051zbypkx0j0td"),
   entityId: z.string(),
   kind: ConversationResponseKind$outboundSchema,
   displayName: z.string(),