+++ packages/orq-rc/RELEASES.md
@@ -5579,3 +5579,13 @@ Based on:
 - [typescript v4.1.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.42 - packages/orq-rc
+
+## 2025-12-15 17:23:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.676.1 (2.781.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.43 - packages/orq-rc